+++ changelog.md
@@ -45,6 +45,7 @@ Codion Change Log
 - EntityTablePanel.TableControl and EntityEditPanel.EditControl renamed EntityTablePanelControl and EntityEditPanelControl, merged with KeyboardShortcut, related refactoring.
 - KeyboardShortcutPanel bug fixed, now displays the correct Add mnemonic.
 - EntityEditPanel.configureControls() renamed setupControls() to prevent confusion. EntityTablePanel.configureControls() renamed setupControls() as well, for consistency.
+- EntityEditPanel, EntityTablePanel, EntityPanel, control setup refactored.
 
 ## 0.17.43
 ### is.codion.common.core